+++ src/main/java/dev/lambdaurora/lovely_snails/LovelySnails.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 LambdAurora <email@lambdaurora.dev>
+ * Copyright (c) 2021-2022 LambdAurora <email@lambdaurora.dev>
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as published by
@@ -30,15 +30,13 @@ import net.minecraft.nbt.NbtCompound;
 import net.minecraft.nbt.NbtElement;
 import net.minecraft.nbt.NbtList;
 import net.minecraft.util.Identifier;
-import net.minecraft.util.registry.Registry;
-import net.minecraft.util.registry.RegistryKey;
 import net.minecraft.world.biome.Biome;
 
 /**
  * Represents the Lovely Snails mod.
  *
  * @author LambdAurora
- * @version 1.0.0
+ * @version 1.0.4
  * @since 1.0.0
  */
 public class LovelySnails implements ModInitializer {
@@ -60,10 +58,11 @@ public class LovelySnails implements ModInitializer {
 					});
 				});
 
-		BiomeModifications.addSpawn(BiomeSelectors.categories(Biome.Category.SWAMP),
+		BiomeModifications.addSpawn(BiomeSelectors.categories(Biome.Category.SWAMP)
+						.or(BiomeSelectors.tag(LovelySnailsRegistry.SNAIL_SWAMP_LIKE_SPAWN_BIOMES)),
 				SpawnGroup.CREATURE, LovelySnailsRegistry.SNAIL_ENTITY_TYPE, 10, 1, 3);
 		BiomeModifications.addSpawn(BiomeSelectors.categories(Biome.Category.MUSHROOM)
-						.or(BiomeSelectors.includeByKey(RegistryKey.of(Registry.BIOME_KEY, new Identifier("dark_forest")))),
+						.or(BiomeSelectors.tag(LovelySnailsRegistry.SNAIL_REGULAR_SPAWN_BIOMES)),
 				SpawnGroup.CREATURE, LovelySnailsRegistry.SNAIL_ENTITY_TYPE, 8, 1, 3);
 	}